backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
daphne
python-dotenv
pillows
django-allauth
